requirements.txt
torch>=1.12
torchvision>=0.13
pytorch-lightning>=1.7
numpy
Pillow
tqdm
scikit-learn